micro-optimization: NULLED vs. return nullptr
Returning a zero state from a native became more complicated, because
if you're in the failure state then nullptr means there was a failure,
but if you're not it's the API version of requesting a null cell
result.

Because it has more processing now, that means disambiguating by
returning the output cell initialized to the nulled cell pattern
isn't as much of a micro-optimization as it was.