Make plain cast() lenient on const pass-thru
There was an idea in the c_cast() construct that the constness you'd
get back would be "merged" between the type you asked for and the
expression that was incoming.

This turns out to be more useful in the general case than giving an
error.  So make it the default, so that the prettier cast() macro
can serve this purpose.