Drop strict_cast() variants
The new methodology for how casts work appears not to trigger the
problems that strict_cast() had to be introduced to workaround.

The explanation is in the commit history if it turns out to be needed
some time in the future, but fortunately it seems the problem has
gone away thanks to the Rewrap mechanic.