Don't handle panic() via Result(T)
When definitional failures were introduced, the fact that they had to
be cooperative with the call stack meant mechanisms were introduced
for that.  These mechanisms could also work for panic, so it was
reasoned that instead of using longjmp, some panics could use the
cooperative mechanism as well.  Such panics would not cause crashes
on systems that didn't support longjmp() or C++ exceptions.

This split panic into two kinds of panics: abrupt and cooperative.

When the Result(T) mechanism was introduced, the idea of letting these
stacks that could pass `fail` states also pass panic states was tried.
However, this was a bad idea... except() couldn't allow you to handle
the divergent failures, but the code path was still running through
it, meaning any allocations the except() was designed to unwind
would not happen.  Basically your system has to have an entirely
distinct bulletproofing mechanism for divergent failures--which Ren-C
has (so long as allocations are done with rebAlloc() and not malloc(),
etc.).

This rips out the bad idea of having handling for divergence in the
Result(T) code, and turns it into a generic hook that just panics.
If there's any handling that's beyond the scope of Result(T) (which
is what RESCUE_SCOPE_IN_CASE_OF_ABRUPT_PANIC() is about.

Also does some upgrading of RESCUE_SCOPE() to work in the usage of
for() loops to scope the error declaration to the block, letting
you choose the name for your error... and also making it possible
to fall through to more code vs. having the scope and catch blocks
forced to be terminal.