panic_abruptly() => panic() ... always abrupt
The concept of "cooperative panics" offered a very slight efficiency
gain while having multiple code paths for producing and handling
panics.  The only reason it would have made sense to keep them would
be if we were going to try making all panics cooperative--but this
is a contradiction in terms.

It's now settled that if a platform is going to be able to do something
else besides terminate on an abrupt panic, it needs to have longjmp()
or C++ exceptions.  Given that, unifying panic under one intention
is the obvious right thing to do.