Make guarantee() work in any expression context
There's a fundamental fact that to work in C, the need for macros
like `trap` and `require` expanding into code that has `return`
statements means they can't act purely as expressions.

That rule doesn't apply to guarantee(), however.  It doesn't return.

Make it easier to use, e.g. as:

    return guarantee (...);

The C++ build adds an assertion via a lambda that there was no failure
in the code.