Don't cast-away Result-ness in downcast
The DowncastHolder was willing to Unwrap_Hot() on a contained Result<T>
and give back a DowncastHolder<T> as a proxy, which was correct...the
downcast holder itself could serve as the "hot potato" variation that
would perform downcasting.

This was mistakenly conflated into thinking that constructing a
DowncastHolder with Result<T> could just be a DowncastHolder<T>, which
is *not* the same thing.  That would lose the Result-ness and cease
to denote the possibility of an error that hadn't been triaged.