Manage/Swap generic to Stub, not Flex
The Stub vs. Flex distinction came along gradually over time, and not
everything went directly to the right split.

There are currently no examples of managed stubs that aren't flexes,
but there could be.  It makes it tricky in the type system to talk
about a type that is on two different branches of Flex if the
managing options aren't common across Stub.