Make Phase a "C typeset" (Details, ParamList)
Phase threw a wrench into the type hierarchy which was just sort of
ignored until it couldn't be: A phase can be either a Details* or
a ParamList, but those types aren't on a fully common inheritance
path (they share Base as a base class, but diverge and C++ won't let
you implicitly cast between such divergences).

To make this less painful to validate, Needful provides extensibility
to its ensure(), so although Phase* parameters of functions can't
directly accept VarList* and Details* you can make a macro that
has no runtime cost and does the validation.