Change guarantee() to assume()
ASSUME more correctly captures the concept of not taking action:
e.g. "guarantee" suggests maybe you'd do something to ensure the
guarantee is held.  "I want you to guarantee that..."

This also puts back a guardrail that the C build includes the assert
that after an assume there's no failure.  That guardrail was dropped
in order to try and make assume more friendly to do things like cast
from the outside, but that's now done better with nocast from the
inside of the expression... and it's just not worth it to lose the
assert in C builds.