Rename all the .test.reb files to .test.r
Historical Rebol, and files in the Rebol library, have used the .r
file extension.

There was a push around the time of the open-sourcing of Rebol to
start using `.reb` - with the argument that the R programming language
(which was a contemporary of Rebol, slighly earlier) was kind of the
de-facto standard for the .r file extension.  It would be easier to
get the right syntax coloring in IDEs and not be mistaken for R files
to change the extension to `.reb`

Experience led to .reb just being a more *cluttered* name.  Rebol and
R files are unlikely to meet in the same codebase, and collisions are
inevitable no matter what you pick.  Plus in the age of AI, the choice
of what syntax highlighter to use will likely be much more intelligent.

The .reb file extension never really was pushed out anywhere (R3-Alpha
used .r3 for Rebol3 files).  This goes back to the .r