Result(Zero) [[nodiscard]]
There had been an oversight and the specialization of ResultWrapper
with Zero was discardable, leading to skipped traps/requires.

At the moment the error handling is half-baked, and the subtleties of
failing and managing all the necessary unwinding vs. panic'ing and
letting the system do automatic cleanup are not really addressed.
This is a primordial effort toward getting a system that can react
safely to things like out of memory errors, and it will take time
before that is actually realized.