Safer, faster, cleaner trap/require/assume syntax
The trap/require/assume macros had tried to avoid problems with being
used incorrectly in branches with a "hot potato" extraction method,
that had runtime cost and still didn't catch all the cases.  Also it
meant that ignored results needed a special form (require vs. required)

This simplifies the model by putting the macro outside any assignment,
so instead of:

    Type* var = require(...);

It is written as:

   require (
     Type* var = require(...)
   );

This actually makes it easier to read what the code is doing and
not treat the require like a function call.  It uses a dummy decl
to avoid being used directly in branches in cases like:

   if (condition)
      require(x = whatever());

That case wasn't caught by the overly-complex prior mechanism.

It's a bit weird, but more solid ground and faster/cleaner/simpler.

The decision to put code on its own line separate from the trap or
require or assume is to help not read it as an extra function call,
and keep the actual code or declaration in the flow of the normal
code more or less...