More descriptive names for Extend_Flex()
It wasn't obvious at the callsites when the number of units used in
the flex was being updated.  This meant there were comments to say
that it was, and sometimes in the absence of comments a redundant
update of the used field.

This just makes the names longer and gets rid of the redundant
length assignments.