Fix bug with CONSTABLE use, remove from Needful
The CONSTABLE macros are sophisticated but hard to use, as a bug
with UTF-8 showed... by not using a CONSTABLE() proxying local
variable, the CONST_IF_C parameter was a universal reference and
modifying the callsite.

This fixes that bug, but also takes the mechanism out of Needful as
it's hard to explain and hard to use... so it's really not something
to try to sell to other projects.