Use Nulled Cells vs. Unreadable for states
There was an idea that making cells unreadable would be a good choice
for avoiding the misinterpretation of cells as the wrong type (e.g.
testing Is_Block() when the contents were meaningless).  This was
used in very limited scenarios vs. as a pervasive policy, and it
introduces cells that are not valid.  It may be that "true unset"
is a better choice in some of these situations, but the few cases
were causing problems with validating routines that saw unreadable
sells as corrupt...and it wasn't worth the trouble.  Switch to nulled
and review the question later--including if tests like Is_Block()
should require you to do tests for null first, like in usermode.