Kill off `atom`/`element` variance in naming
It used to be that calling arguments `atom` was functionally helpful
if they were ^META arguments, because you had to know to use lifted
values with them, and it served as a reminder.

Thanks to "Lift the Universe" meta values can now store unstable
antiforms, which normalizes the interface.  With that normalization,
the variance in naming is just an annoyance.