Restore "truthiness of trash"
Truthy trash was implemented, but then backed out of, when trash was
being used to denote the unspecialized state of variables in frames.

With the "true unset" dual state now serving the role, conditional
tests no longer need to use metaform parameters to test the full
range of stalbe values.  This brings back truthy trash, which seems
to be more useful than not.