Atom required for eval out, MAKE OBJECT! errors
Despite the contravariant output constraints existing, not all routines
that did evaluations were targeting Sink(Atom*)... just atom.  This
means some callsites were passing in Value* or Element*.

Was ignoring the result of evaluations.  This was all right when the
default of unhandled errors was to panic, but now ERROR! comes back
from evaluation normally.

Also, MAKE OBJECT! was ignoring the return result, assuming errors
were handled for it... but that's no longer done.