Fix SWITCH2 demo
Needed to use UNSET? instead of trash? to find slots to fill, now
that trash is no longer used for unspecialized slots.

Before (^result: ...) assignments were legal directly, this used
([^result]: ...), but the change of VOID to being an empty pack makes
the old form not work for VOID, because it tries to unpack an empty
array.  So interesting edge case there, where the new form works and
the workaround no longer does.

Went ahead and switched away from using YES/NO style booleans to just
using okay and null, because it's less hassle--yes and no should only
probably be used when you're trying to put things in blocks and
dialect them.