Updated input position in UPARSE is remainder
This switches from using a separate REMAINDER synthetic variable to
just reacting to changes to the value of INPUT in the combinator.

INPUT becomes packed in as part of the result with RETURN, so it
goes in the output position indicating how much progress was made.

This shortens and simplifies most combinators, in particular those
that don't do any update to the input position.