Fix CELL_MASK_VARLIST check in %d-gc.c
Currently the flag for whether a node gets marked or not has to be
kept in sync with whether that field (if payload) holds a nullptr
or not.  This change was done for efficiency--as every little bit that
makes the GC faster helps.

This means a check in the debug marking was out of date, expecting
the flag in the cell to be set even when the pointer was null.