GC protect g_failure, break Recycle() in sections
The addition of the ability of Executors to return fail through the
same mechanic as action dispatchers meant that g_failure could be
set when a GC ran.  For now, this tolerates that, and just guards
the error.

Also splits up Recycle_Core() into sections using labels, like the
codebase as general is standardizing on.