IF returns NULL, not VOID changes (-> WHEN)
While it was long hoped that an IF that didn't take a branch could
return void (and still have the correct interop with ELSE and THEN),
this hope was ultimately squashed when VOID was resolved as the
unstable antiform of empty pack.

What it would take for a branch to evaluate to void but still run
a THEN would mean putting an empty PACK! in a PACK!, which would not
be naturally decayable...throwing a wrench into the idea.

Hence, WHEN has to be used (or OPT IF or ? IF, though that has
slightly different semantics).