./foo to run member functions, not just .foo
The concept of how .foo works has been changed to make it a synonym
for (.).foo -- this means that having it run functions doesn't work.

To address this, you can just say `./foo`

If this seems to verbose, you can say `use .` at the top of your
member function to pull the definitions into scope.