VOID assignment removes from MAP!
This is based on a realization that if any assignment does removal
from maps (or unsetting of variables) it's best to have that be void,
vs. trying to invent some other way of doing it.

This requires allowing VOIDs to go through assignments more generally,
though it's disabled for object assignments and SET-WORD!s.