Lingering Sequence => Non_Array_Flex
An old use of the term "sequence" was to mean any "REBSER" that was not
a "REBARR".

In the new world, Sequence is a category that encompasses PATH! and
TUPLE! and CHAIN!... interstitially delimited "arrays".

Most terminology had been fixed, but there were lingering references.
Best to just be as literal as possible and call them Non_Array_Flex