Make ADJUNCT name consistent with mainline
Before metaforms or lifting, there was the idea of having an object
that could be associated with an action or a module, which was named
the "meta object".  That name eventually was shuffled to the "adjunct
object" but the bootstrap executable never got synchronized.

This goes ahead and does the change, although the code for adjunct
is far more junky and annoying than it is in mainline...because much
of what adjunct objects were used to do with ACTION! has been absorbed
into the PARAMETER! datatype.