Sync with "Needful" library of mainline
The needful library is breaking out features that were used from the
%c-enhanced.h header file, in a way that can be maintained separate
from the Ren-C codebase.

This adopts that, with main changes to the code being that the Option
type no longer will implicitly convert to integers for switch
statements.  There's also no longer specialized pointer corruption,
but this grafts in a test for whether things are corrupt that has
false positives with nulls.