Resurrect corruption assertion, false positive
The idea that you could assert that a pointer is corrupt was undermined
by the idea that occasionally pointers were corrupted to zero.  It
was deemed that occasionally getting a corrupted pointer as zero was
better than reliably knowing a pointer was a corrupt value.  However,
since there's more than one modality of corruption and also a null
pointer can be accepted as a corrupt state, this switches from being
able to know for sure a pointer is corrupt (Is_Pointer_Corrupted) to
just asserting that it is (Assert_Pointer_Corrupted) which allows
for the false positive, and also means that there's no need to use
an #ifdes around the Is_Pointer_Corrupted calls.