Switch from "enum Reb_Kind" to "Type"
There was a long period of thought about whether or not the word "type"
was a fit for the fundamental byte-oriented type of things, or if
it would be a higher-level concept (e.g. a user might create an
OBJECT! derivation that was a BOOK! and it would have fields like
author etc., and even though it was fundamentally an OBJECT! it would
have a higher level "type").

Reluctance to canonize Type as the word for things like OBJECT! or
BLOCK! led to calling that "Kind" for a long while.  The nuances did
start to creep in which distinguished the Type from the byte known
as the "heart", which eventually merged with the "sigil" to produce
a KIND_BYTE() which ultimately was combined with the LIFT_BYTE() to
calculate something that did get called a "Type" in mainline.

Lacking these subtleties, the KIND_BYTE() just stores a much more
reduced idea of "Type" in the bootstrap executable, so might as well
call it Type and move on.