Tolerate PATH! in GET
You need to be able to write things like (get $lib/append) because it
shouldn't be necessary to convert that into (get $lib.append) if you
are processing code like [lib/append a b] in your dialect.  GET needs
to be able to fetch the action, and confirm that it's an action.