Wrap Scanner_Executor() vs. invasive hook fail()
There was a concept of invasively hooking the fail() macro in order to
let the scanner add special knowledge of where the line and file was
that had the failure.

This takes a cleaner approach, to just have the Scanner_Executor()
wrap a more core function, and if that core function returns a failure
it will label it with the file and line.