Finish name conversion BINARY! => BLOB!
The name was changed in order to differentiate the internal storage
Flex type of Binary from the BLOB! user type, but also because binary
is ambiguous in terms of whether it means 1s and 0s specifically
or things like .EXE and .OBJ are called "binaries".  BLOB! is not
only shorter, but it maps more clearly to what the series type is.