Leverage variadics for angle brackets in macros
There's problems when you mix C++ code with macros with a fixed
number of arguments, because you might use a template expression with
commas in the angle brackets--which is intended conceptually as just
one argument.  The preprocessor doesn't understand that.

However, you can use __VA_ARGS__ macros to work around that.  This
does so, with things like STATIC_ASSERT() and PROBE_CTYPE()