Make port "actor" bodies see containing module
The entire mechanism for defining port actors is completely in question
and just being kept on life support.  For some reason it doesn't go
through normal object creation, but does a dialected interpretation
of the block.  Which it should be able to do, but it makes it more
complicated.  This patches it to use the context of the DEF, not just
the actor object being created.