NULL? doesn't typecheck-pass ~[~null~]~ antiforms
If you say `return: [null?]` in your type spec, that shouldn't pass
heavy nulls... it should decay a heavy null to an ordinary null.

You should have to say [null? heavy-null?] explicitly if you can
return either.

However, this strictness on NULL? means that particular test faces
a dilemma of what to do about heavy nulls.  It can't panic, because
it's a type checker.  Returning false is a bit of a problem also.

So it returns an ERROR! antiform, which causes typechecking to fail
and directs people to say NULL? DECAY if that's what they meant.