Rename `maybe` macro for Option(T) to `opt`
The unwrap prefix macro for Option(T) is intended to be used when you
are certain that you know it's not disengaged.  This is enforced in
debug builds, and some cases might even want to enforce it in release
builds with a runtime check if it were important.

A complement function for extracting called `maybe` was introduced to
do a raw extraction.  This made more sense when the interpreter was
doing what is today OPT as MAYBE (though it wasn't perfectly aligned).
But now that maybe is something totally different related to assignment
it really doesn't make sense.

This commit reluctantly switches over to match with `opt` as the
complement to unwrap, because it's not clear what a better choice
would be...and `maybe` just muddies the water in terms of what
maybe actually does in the evaluator.

Rejected ideas: nowrap (looks too similar to unwrap), unopt (doesn't
read particularly well e.g. as `switch (unopt x)`, and also kind of
suggests 'unoptimized' because we're not used to seeing "unopt"
other places)