Use + instead of << for unwrap/opt/nocast
While << had a correct lower precedence than % but higher precedence
than comparison, Clang warns when you take advantage of this property
in an overload because it thinks you might want to be using the shift
operator in a way that streams, e.g. it would want to stream the
result of a comparison (when you'd not want to shift the result of
a comparison).

Rather than disable the Clang warning or force usage of parentheses
if you write something like (unwrap a == unwrap b) so it would have
to be ((unwrap a) == (unwrap b)), this makes the choice to just use
operator+ instead, which seems to solve the problem well enough.