Add COMMA! handling in CONSTRUCT
CONSTRUCT has weird logic in it, because it doesn't just bind a block
of code and evaluate it... because the rules for binding of the
SET-WORD in the block are different from the rest of the code.

The SET-WORD are bound to the object being created.  The rest of the
code is bound "normally" to the code list being constructed.

This alternative method of evaluation did not account correctly for
the presence of COMMA!