Fix UUID extension
Was doing path mechanics that were unnecessary and causing folders
with the full repo path to be created in the objs dir.

(The mechanic is broken, but fixing it in this bootstrap executable
isn't a good use of time, so just don't create the absolute path).

Also lowercase for os base name.