Panics caught in APIs, free handles
There was a problem with API handles leaking on panic, that's not
entirely clear why it's happening (and not being cleaned up by the
panic).  But given that APIs are targeting returning an error to the
host language layer vs. trying to use the internal exception model,
every API needs to move toward having access to that error.  So
the "May_Panic()" evaluator function is not necessary.

Doing this change happened to fix the problem of the leaked handle.