`make prep` enough to see changes to %prot-xxx.r
Presumably to avoid having to touch the build system, the old method
of building things like %prot-http.r into a C compile-time constant
would make a ".inc" file that would be included in %main.c.  But
this inc file was not a dependency of main.c, so if it changed this
would not trigger a rebuild of main.c

Finding a way to say that main depends on the .inc file would be one
way of addressing the problem.  But nothing else works like that: in
general the files are made as .c and .h ... this uses the same method
by making the size of the file a constant with storage so that even
if main.c isn't recompiled it will work with a new generation of
the temp file.