Primordial GETTER hack, to resurrect %prot-tls.r
When there came to be a difference between obj/x and obj.x - several
"getter" functions in %prot-tls.r stopped being able to have a
comfortable notation where they looked like field access.  Instead
of leaving them as functions, a half-baked implementation of ACCESSOR
was used to allow TUPLE!-based access to a variable whose value
came from code.

This half-baked implementation was soon ripped out, but %prot-tls.r
was left as it was...broken with the idea of being fixed "real soon
now".  But weeks turned into months, and HTTPS reads were broken for
a long time.

To try and put things back on track, this starts with a simple GETTER
that uses FRAME! in the dual band.  It's just a sketch of the mechanic
that's desired, but works well enough to start repairing all the other
things in %prot-tls.r that have broken in the meantime.