Rethink "Undecayable" packs, void assigns (WIP)
The concept of asking if a pack was "undecayable" was a likely not
necessary idea, as opposed to asking to decay and then responding when
you get an error back from that.

This makes that move, and a move toward a more aggressive stance on
decaying errors in packs.  It's a partial step.