Begin unification of OPEN and CONNECT PORT! steps
There was a separation between OPEN and CONNECT due to the idea that
it wasn't always appropriate to connect (e.g. UDP doesn't have a
connection, and if you're acting as a server you are waiting for
clients to connect to you).  This was designed as a way of sharing
network-based OPEN code between connectionless and connected things.

This doesn't really fit the modern way of thinking of connecting
sockets.  Because there can be multiple addresses resolved for a DNS,
and you fold in picking which address to open along with your connect
attempt as part of one operation.

So OPEN needs to know what kind of thing it's opening.  If it needs
a connection, it should do that folded in with the opening...else it
can't actually figure out what the opened IP address is.  But if no
connection is required, it shouldn't open one.