Misc. %prot-tls.r updates for 2025 rules
There was a lapse in maintenance of %prot-tls.r meaning it wasn't kept
all the way up to date with conventions, since it wasn't actually
run.  The code was LOAD-able, and various searches and replaces were
applied to bring it somewhat up to date.  But since it was never run
there were holes.  This patches some of those holes.