Update mbedTLS to 3.6, add ChaCha20 and GCM
The passage of time means cipher suites have moved on and old ones
are deprecated, with example.com only speaking 3 cipher suites NONE
of which were supported by the existing code or crypto:

    TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384 (secp256r1)
    TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256 (secp256r1)
    TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256 (secp256r1)

mbedTLS was last sync'd in May 2022, around version 3.1 -- this steps
that up to version 3.6.

The future directions of mbedTLS in 4.0 have moved to something called
"PSA" (Platform Security Architecture):

  https://mbed-tls.readthedocs.io/en/latest/getting_started/psa/

It's much less granular and transparent than the mbedTLS of yore.  But:

"Mbed TLS 3.6 is a long-term support (LTS) branch. It will be
supported with bug-fixes and security fixes until at least March 2027."