Make DEFINED? its own native
The previous concept of UNDEFINED? was based on the idea that since
GET couldn't return an ERROR! antiform as the thing you were "GET-ing",
it could use ERROR! as the out of band signal that something was
undefined.

Now that GET can actually validly return ERROR! antiforms, this is
no longer seeming like a good idea.  So DEFINED? is made into its own
native entry point.

(This is just to get the WASI build working, it says (get $(1).foo)
is undefined instead of panic-ing, not clear that it should be that
lenient...but this will require more design thought.)