Fix QUIT on command line (adds IGNORE)
PACK! became unwilling to decay when you had ERROR! values in them,
and when that happened there were pieces of the QUIT mechanic that
had not been rigged up correctly to deal with these non-decayable
packs.  Also, there was no function that could defuse the packs,
because they were truly undecayable--they didn't decay to an ERROR!,
they just PANIC'd (decay can't produce an unstable antiform).

This adds IGNORE which is an important tool for throwing away packs
with embedded ERROR!s, given the fact that TRY can't be used.  It
also tweaks some of the broken pieces, like not using `^` to return
an ERROR! in a multi-return.