Generalize ELIDE Semantics
It was noticed that CATCH was glossing over errors e.g. catch [1 / 0]
The current design of CATCH will return NULL if there is no THROW
(this is debated as to whether it's the best answer, vs. letting the
body result be dropped out).

  https://rebol.metaeducation.com/t/catch-return-null/940

So long as that is the design, it's necessary to fail if there are
any ERROR! that are transitively being suppressed.  However, this is
not quite the same as DECAY because no value is desired.  So the
semantics would be the same as those for ELIDE.

This generalizes ELIDE by running it through the same code that DECAY
uses with a flag saying that it's only searching for ERROR!.