Fix typecheck panics for intrinsics
The concept of an intrinsic is that a dispatcher can be called directly
and use a parameter tucked away inside the parent Level's memory, with
no separate Level or Varlist etc.  However, it is responsible for
type checking its arguments (including doing any decaying, or opt-out
features, etc.)

This means that if you call an intrinsic, then it may have to panic
while it is running to report an invalid argument error--which would
usually happen before the dispatcher was called.

However, invalid argument errors during typechecking functions are
not reported--they are ignored and considered to just skip that check.
But a panic() can't be easily undone.

This implements a workaround which sets a flag on intrinsic dispatch
when it's a typecheck.  If that's the case, then instead of giving
an error, it will give the same nullptr result as a bounce as if the
typecheck had failed.  But if the intrinsic is being dispatched in
other cases it goes through error delivery (which has to compensate
for the unusual stack disposition of intrinsics).