Migrate to ^VOID, ^GHOST, ^TRIPWIRE
The idea of having named states being functions had a bad property
that "correct" usage would actually test functions, e.g.:

   >> void? ^void
   == \~null~\  ; antiform

   >> action? ^void
   == \~okay~\  ; antiform

If you had a void variable, you would test it like (void? ^var) and
it would be true.  So it makes sense for the named void state to
be the same.

The impact of this change is meant to be mitigated somewhat by behavior
that makes empty GROUP! e.g. () synthesize a void (voided ghost) by
default.  NOOP synthesizes a tripwire vs. saying ^tripwire, and for
now NIHIL is a function that produces GHOST!