Don't allow Sigil on COMMA! and non-sequencables
When Sigils were introduced [$ ^ @] there are 2-bits of space in every
cell which can be used for them.  However, it turns out they're not a
great idea for every fundamental type, e.g. COMMA!:

    (dollar: '$, at: '@, caret: '^)

If $, is a sigil'd comma, then '$, has to be a quoted sigiled comma,
and that's clearly not what the author meant.