Add TYPECHECK to reduce pressure on MATCH NULL
The idea behind MATCH was to provide typechecking capability and pass
through values that matched, but be null otherwise.  This has a problem
with conflation of nulls that match with a non-matching state of
anything else.

A special mode for MATCH called MATCH:LIFT was used to say you want
the answer lifted, so that a lifted null could be distinguished from
not matching null.  This somewhat strange idea was only used to get
a "yes or no" typechecking answer, which means it makes more sense
to have a TYPECHECK primitive.

This starts that idea out, killing off the :LIFT switch for MATCH.