Get rid of PACK*, plain PACK supports ERROR!
Adding ERROR! support in packs was done reluctantly, but things have
been moving toward where they are legal...but you can't ELIDE them
implicitly or explicitly.

Once that safeguard is in place, restricting the creation of PACK
doesn't make a lot of sense.  So this gets rid of the PACK* special
case routine that was ERROR!-tolerant and just makes that the behavior
of the plain PACK.