Switch from nocast to u_downcast for allocs
nocast is designed as a permissive cast, which will let you do in C++
things that you can do in C.  But this means you can do things like:

    int* ptr = nocast 0;

This isn't really what we want to be using when doing downcasts of
allocations.  That's more of an "unchecked downcast", we want the
right hand side to be a pointer.

So this changes allocation casting to use u_downcast vs. nocast.