Automatically disable int-conversion warning
Gcc and Clang give a warning when programming in C regarding making
things like pointers out of 0.  This is crucial to Needful's workings
in the C language (polymorphic 0).

To make Needful easier to use, this automatically takes care of that
warning in Gcc and Clang via a #pragma.  You can ask it not to do that
if you believe you're using a subset of features which don't require
making pointers from 0, e.g. if you're not using Option() or Result().