NEEDFUL_CPP_ENHANCEMENTS flag
This is a step toward being able to build using Needful in C++ mode
without including the enhanced features.  The goal is to build using
only needful.h without the other parts on disk even if building as C++.

Also renames NEEDFUL_H to NEEDFUL_H_INCLUDED which is more clear.