Make needful.h work in C++ without extra files
The original idea in needful.h was that there would be no C++ in the
header.  But the concept of doing downcasting or converting 0 to an
enum or pointer in C++ is not possible.  Hence things like Option(T)'s
`none` cannot be just #define 0, and nocast or u_downcast cannot be
simply skipped.

This is an attempt to target a very minimal amount of C++ via the
`nocast` operation.  Pursuing this design also leads to the idea of
making a generic concept of zeroness... a Nocast0Struct, which can
be accepted by constructors of user-defined types to participate in
the zero initialization.  While Result0Struct and Option0Struct can
be optionally defined as unique alternatives, this also permits
them to use a minimal Nocast0Struct implementation if you do not
want the ResultWrapper or OptionWrapper classes.

This cleans up a number of issues related to zero initialization,
which was done via hacks previously.