Make Nocast0Struct stateless
Deriving Nocast0Struct from NocastHolder<int> allowed it to reuse the
nocast mechanics, but gave it storage.  As it happens, the conversion
mechanics of nocast are abstracted by NocastConvert and can be used
with a literal 0.

Empty structs can be passed more easily by value, as is standard for
std::nullptr_t, so this makes that change.

Also gets rid of IsOptionWrapper<> as it is no longer needed.