Downgrade HTTPS read in tests to HTTP
We can no longer do reads of `https://example.com` and it's only a
matter of time before other sites become unreadable as well.

Playing catchup with cipher suites is not a project priority (when the
main build to target the public is the web build, anyway).

May be an interesting project for AI to solve right someday.