Try to fix definitional QUIT problems for console
There is an issue where QUIT became definitional, which means that after
a module has finished initializing, any code holding onto its idea of
QUIT that tries to run that QUIT is not allowed to.  The only notion
a module has of QUIT is to say it has finished its initialization
process.

This means the console code itself started having trouble orchestrating
a QUIT using a WORD!.  This rigs it up a little differently just to
get things moving along, and prevent problems with not returning a
1 exit status on CI when commands were failing.