Fixes to LATEST-OF, uses `curl` instead of READ
There were a lot of non-working things in the LATEST-OF script (it
had not been functionally tested for a year).

Scripts no longer synthesize their results implicitly... the last
evaluation doesn't just "fall out"... you have to QUIT:VALUE it.

Checked vs. Debug as a term hadn't been propagated completely.

The debug variant lost its hyphen somewhere.

TUPLE! rendering for numeric tuples now requires you to use @ in the
molding.  It's not clear if this is necessary...perhaps only tuples
and paths and chains which contain blocks or groups are suspect to
render by default?  It doesn't hurt to annotate to say "mold" in any
case, but you might not should have to.

Also, we are no longer maintaining TLS reading...the cost is too high
to bear.  So you have to have curl on your system.  READ and INFO?
may be retrofitted to use curl under the hood, but this is a good
sample of learning how to do that.