Remove import library dependency on Windows
Windows builds actually have not created a r3.lib import library since
the introduction of the RebolApiTableStruct.  This was overlooked since
in the builds of the past year locally on Windows there was a r3.lib
file just hanging around in the build directory to participate in
linkage, if a dynamic extension was used.

Good news is it's not needed now!