Push through OSX => macOS (vs MacOS)
Due to people being pedantic and deferent to the mighty marketing of
Apple, you have to say `macOS` with a lowercase M, even, vs. using the
once-fine term "OS X" or "OSX".

It's us who have to change or look "foolish" instead of them looking
foolish for creating capitalization nightmares ("what happens if you
start a sentence with macOS").

Think different.  And then think about all the annoying people who
will complain if you don't fall in line.  Anyway, since I had gone
though and bothered to change it to "MacOS", this goes ahead and
changes it to "macOS" where case is being heeded...and fixes the
TO_OSX define checks which had not been adjusted (it meant the UUID
extension was breaking, since it thought there was no implementation)