Get rid of "strip" functionality in rebmake
There was a bunch of logic checking to see if you asked for debug
symbols, and if you didn't ask for them it would call strip.  But
if you didn't ask the compiler for debug symbols it should not have
added them in the first place.

Arbitrarily trying to strip symbols gets you errors, e.g. on macOS
if you use system APIs:

    strip libr3-uuid.dylib
    error: symbols referenced by indirect symbol table entries that
    can't be stripped in: libr3-uuid.dylib
    _CFRelease
    _CFUUIDCreate
    _CFUUIDGetUUIDBytes

This downgrades the debug request in debug builds to "asserts", so
it makes the smaller executables by default...which is better anyway.

Changes it so CI scripts do the stripping if they want to before
uploading, and this is more obvious and easier to change.