Use MAKE not TO for string to int in gcc tests
Research is ongoing into the difference between MAKE and TO.  One
idea of TO is that it is reversible, e.g. ("0" <=> 0) for converting
text to an integer.  But if you allow ("00" => 0) then it's not
reversible.  So a theory is that TO would not allow leading zeros,
while MAKE might be more forgiving.  This needs more thought, but
debug builds will panic if a TO transform doesn't reverse.